Factor remote endpoint clients out of remote viewsets
